refactor(audit): split ChangefeedConsumerService into drain + redaction (max-params 2/3)
ChangefeedConsumerService injected 5 deps (surreal, apiKeys, config,
metrics, lease). Split the per-tenant drain engine — SHOW CHANGES →
audit_event translation, batch caps, cursor management — into
ChangefeedDrainService (surreal, config, metrics), and move the PII
redaction allowlist into its own pure changefeed-redaction.ts. The
consumer keeps the cron tick, leader election, tenant fan-out, and
operator status (apiKeys, drain, lease), 3 deps, reading drain.enabled /
drain.sources / drain.perBatchLimit for its status snapshot.

Public surface (tick/stats/drainNow/cursorState) unchanged; the drain
unit-spec now targets ChangefeedDrainService.consumeForTenant and the
redaction spec imports from changefeed-redaction.

Part of the max-params=3 god-class split program.
