+++ src/audit/audit.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { AuthModule } from '../auth/auth.module';
 import { SurrealModule } from '../db/surreal.module';
 import { ChangefeedConsumerService } from './changefeed-consumer.service';
+import { ChangefeedDrainService } from './changefeed-drain.service';
 
 /**
  * Audit module — owns the CHANGEFEED consumer (migration 0023). Kept
@@ -10,7 +11,7 @@ import { ChangefeedConsumerService } from './changefeed-consumer.service';
  */
 @Module({
   imports: [AuthModule, SurrealModule],
-  providers: [ChangefeedConsumerService],
+  providers: [ChangefeedConsumerService, ChangefeedDrainService],
   exports: [ChangefeedConsumerService],
 })
 export class AuditModule {}
